Make track specification required in order to publish a session
Sessions without tracks will now not be listed in session lists, details
or on the schedule until given a track. This makes the handling in the
templates a lot more predictable as there will always be a track.

Conferences that don't use tracks can just create a simple default track
and assign all sessions to it.