Make twitter validator at least do sometihng
Twitter broke our way of doing validation by no longer returning 404
when you query for a profile that doesn't exist, instead returnig 200
with a dynamically loaded error message that's really hard to detect.

The proper way is to use the twitter APIs to do the validation, but that
will require a bit more work. As a stop-gap fix, at least use a regexp
to validate that the characters entered are valid for a twitter name,
thereby  avoiding things like copy/pasted URLs instead of twitter names.