Fix adding additional options with /self/ in reg url
The registration form can be served both as .../register/ and
.../register/self/. The form for upselling additional options would only
work on the former URL, and return 404 on the other one. This would
normally never happen when links were generated inside the system,
because it would never link to /self/ for a finished registration, but
when linking directly to this page from the outside (to avoid the second
step in the user workflow), it could still happen -- and it should of
course not crash.

There are likely more such cases around, but this is the quick-fix.