Add global speaker editor and make it possible to add speakers in backend
This makes it possible to create a speaker profile, not just edit them.
When creating a speaker "inside" a conference, a session must be picked
initially or the record would immediately become unavailable since
speakers without a session in the conference would get filtered.

The global view of speakers requires superuser access as it can mess
with data for multiple conferences.