Don't try to send notificaitons to speakers with no user
If the speaker has no user account, we'd crash. Oops.

We *do* set the notification to "sent" the same way as if they had a
user account, otherwise we can end up never being able to clean out the
queue. Speakers without a user are always going to be special cases, and
they won't be able to go through the verification processes etc anyway.