Fix accouting for overpaid invoices
When the invoie is overpaid we have to put the overpayment *somewhere*
in the accounting, otherwise the accounting entry doesn't balance and an
exception in thrown.

So put this overpayment in as a "negative cost" on whatever account gets
the payment fees, set the description to indicate what it was, and then
leave the accounting entry open for the treasurer to handle.