Don't show time on due *date* field for main site invoices
When an invoice was rendered in the jinja templates of a conference, it
would show the date only (which would make sense for "due date"), but
when rendered on the main templates (such as an invoice for prepaid
vouchers or an invoice clicked through witht he non-logged-in link) we
would show both date and time.

The PDF would also still show just the date.

Given the field name is due date, we should show it as a date...