Properly account for pending additional orders no longer pending
Once an additional order is paid, it's still part of the "pending
additional orders" set (silly naming, yes), but has payconfirmedat set
to a not-null value.

The registration form incrrectly included these attendees twice --
because once the payment is confirmed the additional option is added to
the attendee directly as well. The calculation was done correctly when
showing it to the user, which made it even more confusing.

The backend view of an additional option should also not show the same
attendees both as registered and as "pending but with status paid and
confirmed".