Fix payment statitics query for hybrid bulk pays
If a bulk payment is done where *some* attendees use vouchers and some
pay, the ones with a voucher should not be counted against that invoice.
If they are, then they will either be counted double (if it was a paid
voucher) or incorrectly counted as paid (if it was a free voucher).