Move twitter bot API calls to the background process
Normal notifications were already handled in a background process
(triggered by a NOTIFY to act quickly), so move the API calls for the
bot used for social media posting there as well. This avoids locking up
the UI and/or a webhook in case the Telegram API is unresponsive
(unusual, but can happen).

To do this, abuse the idea of a "channel" to create one called
__telegram_internal, which will hold complete json structures with the
different API parameters we may make.