Re-implement the additional options removal for canceled invoices
This was previously built to use the generic function for expiring
additional options. But since the cancelation of the invoice itself
would update the lastmodified field on the registration, this would
never work -- the expiry would be based on the current time instead,
mening it would never trigger.

Instead do the expiry directly in the invoice cancelation handler. As a
bonus, this means the attendee doesn't get multiple emails saying they
were canceled, just one.