Fix unregistered speaker report for sponsor track speakers
Since sponsor track speakers don't have an account they'd never show up
in the unregistered speaker report, since it matches on id (sidenote:
the button would still highlight as that one did not join with the
registration table).

With this update, show *all* sponsor track speakers in the report as we
don't know, but also add a column that says if we've got a match on the
full name on the speaker which indicates they're *probably* registered.

We should update this in the future to collect the email address on the
sponsor track submission page and match on that as well, which while
still not perfect would give a better match.