Dynamically collapse the list of sponsor emails if long
We already collapsed the list of attendee emails when it got to 5 or
more emails, but did not do it for sponsor emails. It actually makes
more sense to do it there, as it goes in the middle of a shared page
unlike the attendee emails. So copy/paste the solution over from the
attendee emails to have it at both.