Re-implement waitlist confirmation step
Doing this in-line with the waitlist listing becomes very slow and
scroll-intensive when the waitlist is long. So replace that with a
dedicated confirm step that shows just the selected registrations and
the time set, and then issues offers on those.

This also reverts the waitlist pagination back down to 20, undoing the
changes in 9c21654d.