Add "sponsorship benefit reports"
This is a very simple set of reports, but makes it possible to get a
list of benefits across sponsors and levels for printing, including
submitted details when available. Specifically intended to handle for
example "sponsor giveaway" claims where the sponsor has specified what
the giveaway is in the list -- this will allow printing of a list of all
such claims.