Allow manual refunds on automated invoices
Normally this should never be done, as for example a refund on a
conference registration should be done from inside the confreg system.
However, it can sometimes be necessary to override this automation, for
example if a mistake has been made and an incorrect amount refunded, or
if a refund should be done *without* triggering a connected action.