Fix concurrencly safety when posting to social media
The code for sending notifications was properly concurrency safe, but
the one for posting tweets/mastodon/whatnot was not. The individual
cronjob used an advisory lock to avoid running more than once, but if
both the daemon and the cronjob were enabled (the cronjob intended to
catch cases of the daemon temporarily being offline), there was no
proper checking. This could lead to double-posts if the actual posting
took a bit of time, which would typically be the case when large images
were attached to tweets, especially if more than one was approved within
a short interval.

Handle this the same way as we do for notifications, which is wrap it in
a transaction and use SELECT FOR UPDATE.