Stop using hard wrapping for email text areas
This is buggy in both chrome and safari, and most likely other browsers
based on the same heritage as well, in one way (which could lead to a
lot of random blank lines eerywhere in emails). . And has been partially
buggy in Firefox in the past as well. Bottom line is that it does not
appear to be a good thing to rely on, so fall back to juat handling it
manually (until we have managed to add support for html mails).