Make the location of JS and CSS assets configurable
There is a mix of things loaded off CDNs and things loaded from the
local media directory, which is, ahem, inconsistent. This first step
makes the assets a configurable set that can be overridden in
local_settings.py. It does not at this point change any assets, so the
result remains inconsistent, but that will happen in a future update.

Reviewed by Daniel Gustafsson, along with the associated smaller commits