Protect against erasing a version controlled directory
deploystatic.py will erase the unknown files in destpath. It is a wanted
feature which when used carelessly can lead to disaster. While one can not
protect from all disaster cases, a low hanging fruit is figuring out if destpath
is source controlled.

This is admittedly a minor gain, however it also minor to check.