Change format of QR tokens
Instaed of the old format (ID$<token>$ID and AT$<token>$AT), generate a
full URL on the site (using sitebase) in the /t/ namespace. The actual
token is the same, but since it's now an URL it can be reached directly.
A follow-up commit will add an interface using this, but for now it's
just a token format change and an update to the existing scanning paths
to use it.

Both the old and the new token format is supported by the scanner apps.
At some point in the future we can remove support for the old token
format as we no longer generate them, but we will need to leave it
around for a short while to handle overlapping conferences that have
already issued tickets.

Work in progress of #98, but needs the follow-up commit(s) as well.