Allow native token URLs for checkin and badge scanning
With the new tokens that are full URLs, make it possible to scan those
tokens directly using the browser functionality in the phone instead of
using the scanner app (native or web).

Unlike the scanner app, this requires the user (the person doing the
scanning) to be logged in using their community account on the site and
can thus be a bit harder to get started with for those that only use
their account on the computer. It also means an account must now be tied
to the registration of the scanner to work (this was already required
for sponsor badge scanning, but not for checkin processors) for the
native URLs to work (scanner app works like before).

This should hopefully make it easier to work particularly on iOS devices
where Apple changes the behaviour of the web APIs randomly, breaking or
confusing the scanner app. The native Android app is likely still the
fastest and best option for Android users, but this increases the
flexibility for iOS users.

Further work on #98, not quite done yet.