Fix "already stored" message on badge scanning
Commit 6013254d7 changed it so that when a badge was scanned the result
was immediately stored, and only later updated to store the note. This
was done so that sponsors who didn't realize they had to hit the "save"
button wouldn't miss out on their scans. An unfortunate side-effect of
this was that *all* scans of a badge would return the "this badge has
already been scanned" message, even for the first scan. Oops.

Fix this by intoducing a "first scan" boolean variable on the scan which
is set when the GET request is made and reset once the save button is
pressed.

Also reset the state if the user returns to the GET without having hit
save once, as that likely represents a new scan later. This can lead to
false positives in the case of intermittent network issues for example
(if the user just reloaded the browser), but since those would require
the user to take specific action they are likely to be more
understanding of what happened.