Extend the news json feed to also include social media posts
By specifying ?parts=news,posts it's possible to get both parts, and by
just ?posts to get, well, just the posts. If nothing at all is
specified, the news posts are returned, and they are returned in a
backwards compatible format (with no top level key in the json). To
return just the news in the new format, specify ?parts=news.

Putting both these (optionally) behind a single endpoint can reduce
the number of http requests for the common case of a site wanting both.