Send a list of existing sponsor emails to newly confirmed sponsors
When a sponsor is confirmed after emails have already been sent to the
level, the previous emails are available on the sponsorhsip dashboard,
but commonly the sponsor will not realize this and read them. To help
with this, send an explicit list of the emails that have previously been
sent including a link to view each and every one of them. This email is
sent as soon as a sponsor is confirmed, provided there were any previous
emails sent.

Fixes #87