Don't accidentally put back all providers when sending fails
If a tweet was queued for multiple providers and one of them failed,
then the save() trigger in ConferenceTweetQueue would accidentally put
back *all* providers on the list of what remains to be done, and thus
would re-post to the successful ones. If all postings were successful,
then the sent attrbute would be set to True (by the sender), and it
would not re-populate the list of remaining.

Make sure we only update the list of what remains when it's empty --
that way it will only fire on a new entry. When it's empty because
everything has been sent then "sent" is already set to True, so we end
up not adding it back already.

Reported by Andreas Scherbaum. Debugging help from Daniel Gustafsson.