Sleep 60 seconds instead of 1 if there is a posting error
This is an ugly quick-fix and we need to put in place a better handling
of backing off. Right now if there is a permanent failure on a post, we
will get stuck trying that same post every 60 seconds until it succeeds
without letting anything else past. We need to figure out a better
long-term solution.