Override the git home directory if requested
If the calling environment requests it (by setting
PGEUSYS_GIT_SET_HOME), override HOME when calling git operations.

This is needed for some webservers with newer versions of git, which
require a safe directory to be set, but uses $HOME to find the home
directory rather than looking it up. Many webservers won't populate HOME
(for security reasons, probably), so inject it specifically for the git
operation when needed.