Deactivate the talk vote href link.
Talkvote is build based on table. The cell on the status column for each talk,
is loaded with an onclick action, which fires up a js function. It also contains
a href link, targeting an empty url fragment. The purpose of that link is
deduced that is mostly aesthetic, in order to give to the enclosed text the
appearance of a link. However, an empty fragment, will cause the page to roll
back to the top, if the current view window has scrolled away from it. The
combination of the two, i.e. the onclick on the whole cell, and the href target
is antagonistic.

Combat it, by deactivating the href link with a false onclick action. This
assures that aesthetically the page remains unchanged and that clicking on the
link will not cause any unwanted scroll.