Improve experience when changing status on voting page
One of the ways to change the status of a submission is to click on the
submission's status link. Then a pop up dialogue appears. However the placement
of the pop up is not relative to the submission. It may be end up being away
from the current viewport, based on the browser used and the length of the page.

Mitigate by explicitly placing the dialogue relative to the clicked event. While
at it, add the submission's id on the dialogue's title so that there may not be
any confusion as to which submission is being edited.