Centralize registration of messaging accounts to registrations
This code was basically duplicated between Twitter and Telegram, so
let's centralize it also in preparation for future providers.

Also add a reglog() call when an account is associaed with a new
messaging.

In doing so, refactor some functions out into util/messaging/common.py
to avoid circular dependencies on import.