Don't crash on in-flight regs when viewing addopts and discounts
The lists for discount codes and additional options in the backend shows
which registration type the different regs that use them have. This will
always be set once a registration is finalized, but during creation the
regtype can be empty (unless the addopt or code requires specific
regtypes in which case this can't happen), and we still list those on
the pages.

So handle that by just rendering the regtype empty if there is none
picked.