Retry with truncated text if twitter posts are too long
Since we can't exactly reproduce how twitter calculates how long a text
is (it's not documented...), we can sometimes end up submitting a text
that's too long even after we've truncated it. In this case, try to
identify the explicit error code that's given, truncate the text by one
word, and try again. Keep trying until it goes through (or until we're
out of words) with a 1 second gap between.