Improve wording on admin signup notification
The text was slightly unclear, as it announced an attendee having signed up even
in the cases where the attendee had chosen to not do so.

Given that the value of 'attendeesignup.choice' is a free text and not encoded
in the system, the text can not be controlled. A generic phrase was chosen
instead.