Add ability to preview/generate badges for multiple attendees
Previously this could be done either through clicking the Preview badge
on each individual badge, or through genereating a report (where,
depending on the situation, it might be difficult to set the criteria).
This plugs the "hole" in between them by also making it possible to
select multiple attendees in the registration list and generate badges
specifically for them.

We do explicitly allow the generation of badges for registrations not
yet canceled, as one might want to pre-emptively print those.