Add support for NULL filtering in attendee reports
Previously we only had the ability to filter for NOT NULL, but for
example when printing badges and attendee lists it makes sense to also
be able to filter for NULL (canceledat IS NULL is what one would
normally want there). Do so by explicitly putting "\N" in the field.