Don't crash if there are no channels at all on a telegram provider
When trying to notify about twitter  moderation, we would correctly exit
with no issues if there was a telegram provider configured but it had no
channel for social media management configured. But if it had no
channels *at all* configured, we'd crash. Fix that for an early exit in
this case as well.