Properly unlink prepaid voucher on invoice cancellation
Normally an invoice with a prepaid voucher is never generated, but if
the user *also* adds an additional option, it can be. And if this
invoice was then canceled (or expired), the voucher field would be
cleared from the registration, but not actually unlinked from the user
and therefor listed as used if viewed from the voucher list, and not
reusable.

It was correctly unlinked if a bulk payment was used, but not for an
individual registration. So copy the method from the bulk invoice
processor.