Give better error message for unregistered sponsors
Both the case of a user not being registered at all and a user being
registered but not added as a badge scanner would just generate a
generic 404 message, making it unclear what the state was. Make both
those error messages clearer by telling the user exactly what is wrong.

In passing also verify that the user is actually confirmed registered
and not canceled. Previously that state would go forward to the next
step and complain that the user wasn't a *badge scanner*, even if they
weren't confirmed. Better let them know the full state up front.