List sponsorships to manage on registration dashboard
When a users is a manager of a sponsorship, list it on their dashboard
as this is the typical starting point of any entry. We'd already list if
they were badge scanners, but let's also do it for sponsorships.

List each individual sponsorship with a separate link, as it's highly
unlikely outside of dev environments that somebody has many of them.