Add support for asking for recording consent in cfp
This adds an optional checkbox to the cfp form, asking for consent to
record the session (enabled/disabled on a per conference level). The
status of this field will be reflected on the talk voting page, the
schedule creation page and will be available to the schedule creation
templates if needed.