Disallow a regular attendee reg to be part of their own multireg
If an attendee created a registration for themselves, saved it but
didn't pay for it and then navigated to the multireg page their own
registration would show up there. If they then edited it *from there*,
it would get disconnected from their account. This could cause issues
for example if that registration was on the waitlist, because once the
account was disconnected there would be no way to accept the offer on
the waitlist. Other weird states could most likely also happen given
it's a "shouldn't happen" scenario.

As this is a state the system is not expecting to handle, prevent it. If
the user wants their own registration to be part of a multireg they can
still do that, they just have to create it in the multireg in the first
place, it cannot be transferred. (They then get to connect it to their
account the same way as everybody else, once the multireg is paid).