Add tools and data to diff names of constraints created by django
The names of constraints (unique or foreign key) are *not* predictable
between versions. This means that the name of the constraint in the
database will depend on which version of django was used when the
constraint was *applied*, not when it was created - meaning that systems
installed at different times will have different names for the
constraints.

This can cause future migrations to fail because django tries to ALTER
TABLE or DROP CONSTRAINT or such things using the name it thinks the
constraint has, but because of the version used it might have a
different name. Upgrading django does *not* create any form of
compatibility mgiration, so you're basically left with a database that
can't be migrated, unless all migrations are always applied with the
same version. Which is of course not the case when for example a brand
new installation is made.

To work around this limitation, we dump a list of all unique and foreign
key constraints in the database to files, and use those to compare. The
assumption for now is that django will only change the names in major
versions, which means we only need to re-dump these files in major
versions. They should be dumped using a fresh install in an empty
database, and can then be used to backfill old databases with new names.

A command to diff and automatically rename constraints is also included.
Once proven to work it will also be turned into a migration step so it
can be automated properly, but let's verify first. This command is only
capable of renaming incorrectly renamed ones - if there are constraints
completely missing or added that shouldn't be there, it will print the
results but not do anything about it.