Add a field for "invoice extra description", use in confsponsor
This is a free text field added to each invoice (optional). If
specified, the text in question will be included on the invoice payment
page and in the email sent to the invoice recipient (if an email is
sent).

Also add a field to the SponsorshipLevel model in confsponsor, that
allows you to specify this individually for each sponsorship level. It
will then be passed through to the invoice and included everywhere
(sponsorship invoices are always automatically emailed). This is
intended to for example inform sponsors about optional other payment
methods or oher such information.