Implement Abstain votes in the session voting system
This vote differs from No Vote in that it gets marked as a vote (and is
no longer highlighted in red). It is still not counted against the
average for the session.

This status is intended for cases when one is avoiding to vote on
sessions submitted by a colleague or oneself for example.

Fixes #126
Fixes #127