Actually delete prepaid vouchers when benefit is unclaimed
As before, if any vouchers have been used, a benefit cannot be
unclaimed. But previously if no vouchers had been used the benefit would
be unclaimed but the vouchers remain, which clearly was wrong.

Instead, implement a process_unclaim callback that is called if a
benefit is confirmed but yet allows unclaiming. It should basically undo
whatever process_confirm did, and for now that only applies to entry
vouchers.