Make all sponsorship benefits potentially autoconfirmable
Previously some would always autoconfirm (specifically, entry
vouchers), some made it configurable ("requires explicit claiming"),
and some made it impossible. This is now moved to central
functionality so all types of benefits can have it. The old state is
maintained by materializing the old defaults onto existing benefits.

In doing so, attendee voucher processing is moved to process_confirm()
where it should really already have been. This was the only benefit
that took any persistent actions in the save_form() method, a leftover
from old times.

It's configurable on each benefit class if it can be autoconfirmed or
not, but at this point all available classes support it.

In passing, centralize the "decline benefit" functionality as much as
possible (it can't be 100% central as it looks slightly different for
different benefit classes, but the "execution" of the decline can be
and now is), removing a lot of dupliated code.