Restructure query and building of session voting page
Instead of having some group by done in the templating and some html
building done in the SQL, separate them apart so data processing is in
SQL and HTML generation in the templates... Also removes a very ugly
"transform" (semi-group-by) that was done in the python layer. Leads to
a much more readable query (and faster, as a bonus), but also one that
is less fragile for future changes.