Support inheriting jinja templates with the same name
This allows for example confreg/regform.html in a conference skin to
inherit from confreg/regform.html in the base system, and replacing only
parts of it. Previously this would lead to infinite recursion and not
work unless a different name was used, and if a different names was used
the system wouldn't find the template to load it.

The implementation basically "pops the directory off the stack that is
the search path" if it detects a template trying to load another
template with the same name, and then repeats that if necessary.

Most, if not all, templates aren't really set up to have parts of them
inherited other than the ability to replace the <title> tag, but with
this foundation being available we can start expanding on that in the
future.