Add a jinja filter to look up a context variable by name-in-variable
This allows code to store the *name* of a variable in a variable, where
the template can then look it up in the gloabal context for the page.
This will allow, among other things, a conference template context (a
concept that didn't exist back when this was originall created, but does
now) to override a text string that cannot be handled as a subtemplate.

As with the previous commit, the current code and templates aren't fully
prepared for this yet, but they can be adjusted with it as needed.
(There are some use cases that already work, but the majority won't)