Move cfpintro, additionalintro and welcomemail from db to templates
This moves these large-text fields out of the database and into the
templates, which makes them a lot easier to edit there, and will make it
possible to use better HTML templating and such in the future.

cfpintro can be set either using a full template block override or using
a property called system.cfp.intro.
additionalintro can for now only be set as a property,
system.reg.additionalintro.

In passing also add a system.reg.voucherintro to override the default
text introducing what to do with vouchers.

As of this, require the template confreg/mail/welcomemail.txt to exist
in the skin directory to make it possible to send welcome emails.
Previously there was a default content for this file that just included
the database property as the entire text, but from this it requires a
complete file in the skin. The fact that it exists is validated when
trying to turn on either welcome email or tickets.

Also create a new model for ConferenceRemovedData where we persist the
current values for these fields, since the database content isn't under
version control, in case it needs to be recovered later.