BREAKING CHANGE: Allow the currency format to be configurable.
This change allows the currency format to be configured in the
instance. This allows us to use alternate symbols (e.g. £), and
to format amounts using the format required in different
jurisdictions.

The currency format is set to Euro by default.

Note that this change requires an update to any Jinja templates
that use the currency_format tag. This must be changed to
format_currency.