Split attendee report process into two steps
This has one step for filter+fields, and another step for generating the
actual report. This allows the user to experiment with different filters
to find the right ones without losing what's there now, compared to the
current setup where everytime you generate a report you have to start
over if you get it wrong.

This also sets up a foundation for some further additions to do more
flexible reporting based on more dynamic properties, but none of that is
built yet.