Re-factor advanced attendee reports to be encapsulated
Encapsulate the management of advanced attendee reports in a class which
can cache things like the list of filters and the list of fields. This
avoids some repeated lookups of the same thing, and gives for a cleaner
implementation. It also lays some groundwork for some more dynamic
features int he advanced reports coming in a later commit.