Support deleting keys in json_form_fields
If the attribute delete_on_empty is set to True on the Field, and the
value submitted from the user is empty, delete the key from the json
object instead of setting it to the empty string. The default behaviour
remains the previous one, which is set the key explicitly to the empty
string.