Implement dynamic properties on attendee registrations
This allows each conference to define zero or more dynamic properties
for attendees, each which can hold a string value. These properties can
be used to filter reports, same as any other string value. Storage in
the system is backed by a jsonb field.

This can be used to for example track which registrations have had their
badges printed, by declaring a field that holds a "badge batch" and then
filter based on that.