Show which track a session is confirmed in when it's confirmed
If there was a choice of tracks available when the speaker submitted,
also confirm to them which track they were accepted in, as this is
somethign that may have been changed by the organisers (pgeu uses tracks
for example for FOSDEM to separate out the different days).

If tracks are not choosable in the cfp form, then we don't change the
notification.

Fixes #5