Restructure the check for conference in jinja sandbox
The old code would cause a lot of extra lookups of the conference that
resulted in many queries to the database unnecessarily. So instead of
doing that, rely on the djanog _meta api to figure out which fields are
there which can be done without triggering lookups. This drastically
reduces the number of queries to render the registration dashboard for
example (and less drastically reduces it on many other pages as well).

Reviewed by Daniel Gustafsson