Explicitly include the conference date in the invoice title
Including the event date on the invoice is required in some
jurisdictions and it's easiest to just do it in the title.

In passing, adjust it so the title of single registration invoices
created directly and single registration invoices created through
the "register for somebody else" are now the same. This removes
the email address from the invoice title of a register-yourself
attendee. But since it's still included in an invoice row, it should
not matter (I believe having it in the title in the first place dates
back to when we didn't have invoice rows).