Remove "all available feedback" view
This was presend on /events/feedback/ but not linked anywhere. And it
makes little sense to list a bunch of old conferences long in the past,
just because nobody explicitly closed the feedback. The normal place to
link to the feedback is the per-conference pages, which then links
directly to the feedback for that specific conference.