Fix headers in talkvote page
When moving to the new query in commit 93251acf, the "current user" data
is not always in the first column. The part of the template that shows
the user voting in the *header* of the table did not get the memo about
that, and thus would show the wrong username. The contents of the table
was correct, but having the wrong username in the header is obviously
quite confusing.

Fixes #129