Add a .git-blame-ignore-revs file
Works the same as we do in the main postgresql repo, allows git blame to
exclude some revisions.

To use, you have to run:
$ git config blame.ignoreRevsFile .git-blame-ignore-revs

There are probably more commits to add to this, but this includes the
major tabs->spaces change which tends to break a lot of git blame, as
well as a few more related to e.g. the 2to3 changes.