Unlink sponsorship contracts before canceling them
If we don't do this, the webhook receiving the "canceled document" event
will crash when it tries to look up the sponsor to reject them. But
we've already deleted it when it was done as part of the rejection.

So we remove the document from the sponsor, and leave the document
itself orphaned so we have traceability.