Don't crash reg dashboard on templates that don't have addoptintro
If additional options existed but no intro was provided in the context,
there'd be an undefined variable crash. All other places like cfp used
the lookup function that handles this case, but not the registration
dashboard. Bug introduced in a1badc02 where it was moved from the
database to the templates.