Re-send policy email instead of welcome email if policy not accepted
The initial code for policy assumed this scenario could never happen and
threw an exception. However, it can happen - when an admin chooses to
send a new welcome email before the policy has been accepted. This is
done for example when a registration was done with an invalid email
address (which can't be done directly, but it can be done when
registereding for somebody else, this being the loophole missed in the
initial commit) that was later fixed.

Instead, just send the policy email, because once the policy is accepted
that will trigger the welcome email automatically.