Teach social media posting to differentiate contents to providers
This allows the core of the system to differentiate and post the same
post with different content to different social networks, for example a
different content to Twitter and Mastodon. This is primarily intended
for things like campaigns where speaker social media handles are
highlighted - the handles are different on different networks, and you
don't want to post a mastodon handle on twitter and vice versa.

The campaign system (which currently only has one type of campaign,
which is the confirmed sessions campaign) is extended to handle this and
be able to highlight the correct handle, using a new jinja filter
introduced.

Speaker profiles are extended to not only collect twitter names, but
instead collect all social networks that are supported on the instance.

The implementation supports the generic idea of "messaging provider"
that we've had for a while, and is thus already ready for if support is
added for more providers in the future. All the old hard-codings to
twitter are removed from speakers and from the core system, but remains
on registrations (where we just collect the twitter name and don't do
anything with it) and sponsors (room for improvement here).

NOTE! This changes the format of the news.json output when social media
posts are included. I've not been able to find anybody actually using
this functionality yet, so I just went with breaking the old format so
as not to have to drag legacy around forever.