Teach the sponsor system about multiple social media providers
This makes the sponsor system work the same way as the speaker system in
that each sponsor gets to register multiple social media handles, one
for each configured type of broadcaster (for now, twitter and/or
mastodon). When generating automatic posts upon confirmation of beneits,
if the social media handle is used in the template, different versions
will post to different networks. The same `social` filter in the jinja
table applies to both sponsors and speakers.